# api/api.yml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Task Management Service
components:
  schemas:
    NewTask:
      type: object
      required:
        - name
        - ownerId
        - estimation
        - reminderPeriod
      properties:
        name:
          type: string
        ownerId:
          type: string
        estimation:
          type: number
          format: float
        reminderPeriod:
          type: number
          format: float

    Task:
      allOf:
        - $ref: '#/components/schemas/NewTask'
        - type: object
          required:
            - id
          properties:
            id:
              type: string

    User:
      type: object
      required:
        - id
        - firstName
        - lastName
        - email
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string

paths:
  /tasks:
    post:
      summary: Create new task
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewTask'
      responses:
        '201':
          description: task is created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
        '400':
          description: Bad request
        '409':
          description: Conflict in insertion
        '500':
          description: Internal server error

    get:
      summary: Get all tasks
      operationId: getTasks
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
        '400':
          description: Bad request
        '500':
          description: Internal server error

  '/tasks/{id}':
    delete:
      summary: delete task
      operationId: deleteTask
      parameters:
        - name: id
          in: path
          description: ID of the task
          schema:
            type: string
          required: true
      responses:
        '201':
          description: task is deleted
        '400':
          description: Bad request
        '404':
          description: Not Found
        '500':
          description: Internal server error
    get:
      summary: Get task by id
      operationId: getTask
      parameters:
        - name: id
          in: path
          description: ID of the user
          schema:
            type: string
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request
        '500':
          description: Internal server error
  /users:
    post:
      summary: Create new user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: user is created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
        '400':
          description: Bad request
        '409':
          description: Conflict in insertion
        '500':
          description: Internal server error
    get:
      summary: Get all Users
      operationId: getUsers
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        '400':
          description: Bad request
        '500':
          description: Internal server error

  /users/{id}:
    delete:
      summary: delete user
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          description: ID of the user
          schema:
            type: string
          required: true
      responses:
        '201':
          description: user is deleted
        '400':
          description: Bad request
        '404':
          description: Not Found
        '500':
          description: Internal server error
    get:
      summary: Get User by id
      operationId: getUser
      parameters:
        - name: id
          in: path
          description: ID of the user
          schema:
            type: string
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request
        '500':
          description: Internal server error